requirements.txt
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
python-dotenv==1.2.1
rich==14.2.0
google-genai